feat(ui): new settings form layout
fix #2285
