+++ renderer/components/Settings/SettingsForm.js
@@ -1,0 +1,50 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import { injectIntl, intlShape } from 'react-intl'
+import { Form } from 'components/UI'
+import messages from './messages'
+
+const SettingsForm = ({
+  intl,
+  configureAutoUpdater,
+  onSubmit,
+  setLocale,
+  saveConfigOverrides,
+  showNotification,
+  children,
+}) => {
+  const handleSubmit = async values => {
+    // Save the updated settings.
+    await saveConfigOverrides(values)
+
+    // Special handling.
+    if (values.locale) {
+      await setLocale(values.locale)
+    }
+    if (values.autoupdate) {
+      configureAutoUpdater(values.autoupdate)
+    }
+
+    // Show a notification.
+    const message = intl.formatMessage({ ...messages.submit_success })
+    showNotification(message)
+
+    // Finally, run any user supplied submit handler.
+    if (onSubmit) {
+      onSubmit(values)
+    }
+  }
+
+  return <Form onSubmit={handleSubmit}>{children}</Form>
+}
+
+SettingsForm.propTypes = {
+  children: PropTypes.node,
+  configureAutoUpdater: PropTypes.func.isRequired,
+  intl: intlShape.isRequired,
+  onSubmit: PropTypes.func,
+  saveConfigOverrides: PropTypes.func.isRequired,
+  setLocale: PropTypes.func.isRequired,
+  showNotification: PropTypes.func.isRequired,
+}
+export default injectIntl(SettingsForm)
